backend/requirements.txt
fastapi
uvicorn
pytesseract
requests
python-multipart
python-dotenv
Pillow